Update for new pwndbg heap classes
Pwndbg uses a virtual environment now too, so install the dependencies inside the venv instead of globally.

The heap analysis code was refactored in pwndbg so do the same in vheap.

Instead of inserting the code into pwndbg, access the api from a standalone file. This makes vheap agnostic to updates to the heap.py file in pwndbg (except api changes of course).

Stop the server when gdb exits too to avoid hanging.

Includes updated dependencies in the frontend from #2 by @rb-x